backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
slowapi==0.1.9
email-validator==2.2.0
pytest==8.3.2
pytest-asyncio==0.23.8
rapidfuzz==3.9.4
python-multipart==0.0.17
